♻️ Move `Vec` out from `config`
